Moved node/src/connect into offst-app. offst-app now does not depend on offst-node.
