+++ components/app/src/app_conn/buyer.rs
