+++ components/app/src/lib.rs
@@ -15,6 +15,7 @@
 #[macro_use]
 extern crate log;
 
+mod app_conn;
 mod connect;
 pub mod gen;
 mod identity;
@@ -31,9 +32,9 @@ pub use proto::funder::signature_buff::verify_receipt;
 pub use proto::index_server::messages::NamedIndexServerAddress;
 pub use proto::report::signature_buff::verify_move_token_hashed_report;
 
-pub use node::connect::{AppBuyer, AppConfig, AppReport, AppRoutes, AppSeller, NodeConnection};
+pub use self::app_conn::{AppBuyer, AppConfig, AppConn, AppReport, AppRoutes, AppSeller};
 
-pub use self::connect::{connect, ConnectError};
+pub use self::connect::{connect, node_connect, ConnectError};
 pub use self::identity::{identity_from_file, IdentityFromFileError};
 
 // TODO: Possibly reduce what we export from report in the future?
