--- components/node/src/connect/mod.rs
@@ -1,8 +1,0 @@
-mod connect;
-mod node_connection;
-
-pub use self::connect::{node_connect, NodeConnection};
-
-pub use self::node_connection::{
-    buyer::AppBuyer, config::AppConfig, report::AppReport, routes::AppRoutes, seller::AppSeller,
-};
