--- components/node/src/connect/node_connection/mod.rs
@@ -1,9 +1,0 @@
-pub mod buyer;
-pub mod config;
-pub mod report;
-pub mod routes;
-pub mod seller;
-
-mod node_connection;
-
-pub use self::node_connection::{NodeConnection, NodeConnectionError, NodeConnectionTuple};
