+++ components/test/Cargo.toml
@@ -19,6 +19,7 @@ node = { path = "../node", version = "0.1.0" , package = "offst-node" }
 database = { path = "../database", version = "0.1.0" , package = "offst-database" }
 bin = { path = "../bin", version = "0.1.0" , package = "offst-bin" }
 stctrl = { path = "../stctrl", version = "0.1.0" , package = "offst-stctrl" }
+app = { path = "../app", version = "0.1.0" , package = "offst-app" }
 
 futures-preview = {version = "0.3.0-alpha.16", features = ["compat"] }
 futures-test-preview = {version = "0.3.0-alpha.16"}
